Split out .cpp file.
